# Add `High Quality` sun sampling mode. Fix issue where enabling draw sun but not enabling sunlight would draw a sun with no contribution to the scene.