--- a/chunky/src/java/se/llbit/chunky/renderer/SunSamplingStrategy.java
+++ b/chunky/src/java/se/llbit/chunky/renderer/SunSamplingStrategy.java
@@ -19,12 +19,12 @@ package se.llbit.chunky.renderer;
 import se.llbit.util.Registerable;
 
 public enum SunSamplingStrategy implements Registerable {
-    Off("Off", "Sun is not sampled with next event estimation.", false, true, false, true),
-    NonLuminous("Non-Luminous", "Sun is drawn on the skybox but it does not contribute to the lighting of the scene.", false, false, false, false),
-    Fast("Fast", "Fast sun sampling algorithm. Lower noise but does not correctly model some visual effects.", true, false, false, false),
-    HighQuality("High Quality", "High quality sun sampling. More noise but correctly models visual effects.", true, true, true, true);
+    OFF("Off", "Sun is not sampled with next event estimation.", false, true, false, true),
+    NON_LUMINOUS("Non-Luminous", "Sun is drawn on the skybox but it does not contribute to the lighting of the scene.", false, false, false, false),
+    FAST("Fast", "Fast sun sampling algorithm. Lower noise but does not correctly model some visual effects.", true, false, false, false),
+    HIGH_QUALITY("High Quality", "High quality sun sampling. More noise but correctly models visual effects such as caustics.", true, true, true, true);
 
-    private final String friendlyName;
+    private final String displayName;
     private final String description;
 
     private final boolean sunSampling;
@@ -32,8 +32,8 @@ public enum SunSamplingStrategy implements Registerable {
     private final boolean strictDirectLight;
     private final boolean sunLuminosity;
 
-    SunSamplingStrategy(String name, String description, boolean sunSampling, boolean diffuseSun, boolean strictDirectLight, boolean sunLuminosity) {
-        this.friendlyName = name;
+    SunSamplingStrategy(String displayName, String description, boolean sunSampling, boolean diffuseSun, boolean strictDirectLight, boolean sunLuminosity) {
+        this.displayName = displayName;
         this.description = description;
 
         this.sunSampling = sunSampling;
@@ -44,7 +44,7 @@ public enum SunSamplingStrategy implements Registerable {
 
     @Override
     public String getName() {
-        return this.friendlyName;
+        return this.displayName;
     }
 
     @Override