--- a/chunky/src/res/se/llbit/chunky/ui/render/tabs/LightingTab.fxml
+++ b/chunky/src/res/se/llbit/chunky/ui/render/tabs/LightingTab.fxml
@@ -23,6 +23,8 @@
     <CheckBox fx:id="enableSunlight" mnemonicParsing="false" text="Enable sunlight" />
     <CheckBox fx:id="drawSun" mnemonicParsing="false" text="Draw sun" />
     <DoubleAdjuster fx:id="sunIntensity" maxWidth="1.7976931348623157E308" />
+    <CheckBox fx:id="fastSunSampling" mnemonicParsing="false" text="Fast Sun Sampling" />
+    <DoubleAdjuster fx:id="trueSunIntensity" maxWidth="Infinity" />
     <AngleAdjuster fx:id="sunAzimuth" />
     <AngleAdjuster fx:id="sunAltitude" />
     <HBox alignment="CENTER_LEFT" spacing="10.0">