Closes #1184.
This adds a `Sun Luminosity` setting which changes the true brightness of the sun when intersected with organically and a new sun sampling mode which does both sun samples and allows for sun intersections for diffuse materials. This new sun sampling mode allows for visual effects like caustics (with much render time):
![image](https://user-images.githubusercontent.com/42661490/162601080-d96f82a0-7a46-41f1-abc3-073bd18b72ac.png)