@leMaik 2022-04-10
Nice! :)

> Fix issue where enabling draw sun but not enabling sunlight would draw a sun with no contribution to the scene.

I always assumed this would be a feature so that one can use sky light and emitters only but still have a visual sun?