@jackjt8 2022-06-11
Compared to `2.5.0-SNAPSHOT.124.ge9d7e123` (obtained today)

Master/current nightly - 441,874 SPS | 46 s

This PR:
- SSS:OFF/NON_LUMINOUS ~~632,919 SPS | 32 s
- SSS:FAST 438,496 SPS | 46 s (approx. matching master)
- SSS:HIGH_QUALITY 497,111 SPS | 41 s

---

I do not see a performance regression from this change.